+++ src/main/java/pl/asie/computronics/Computronics.java
@@ -358,7 +358,7 @@ public class Computronics {
 
 		// Mod compat - GregTech
 		/*if(itemTape != null && Mods.hasVersion(Mods.GregTech, Mods.Versions.GregTech5) && itemPartsGreg != null) { TODO GregTech
-			GregTechRecipes.registerGregTechTapeRecipes();
+			GregTechRecipes.registerStandardGregTechRecipes();
 		}*/
 
 		if(Mods.isLoaded(Mods.OpenComputers)) {